systemProject/OCR/PerfectOCR/requirements.txt
# PerfectOCR — Dependencies
# Core
python-dotenv>=1.0.0
Pillow>=10.0.0
PyMuPDF>=1.23.0         # fitz — PDF to image conversion

# Gemini API
google-genai>=1.0.0

# OpenAI API (GPT-4o)
openai>=1.0.0

# Streamlit Dashboard
streamlit>=1.30.0

# Data display
pandas>=2.0.0

# Fast OCR (Optional - for optimization)
# Requires Tesseract binary: https://github.com/UB-Mannheim/tesseract/wiki
# pytesseract>=0.3.10
